Final Changes to Add Teams Logic.
Just need to look at the front end of the Add Teams page. There is a problem with the reloading of the page which is gonna cause the program to be much Slower. Also I'm struggling to make optimize the HTML code to have the Tables be in the correct divs and because of this I need 2 foreach loops. Will maybe have to revise this.